Capsule Card Component
I replaced references to external assets and added to the project in a
previous commit. Along with splitting parts of this component into
smaller components. The new components I created are for displaying
stats and tags. A lot of animation was also added but will require
minor tweaks which will be added in a future commit.