Pseudo Data and Calculated Discount Price
Updated CapsuleCard and discount module to pass discount and price prop
in order to dynamically calculate discount price along with setting a
style modifier if there is no discount.

I added some dummy data to App.jsx. I'll move this to a dummy file or
maybe I'll just create an API in AWS.