Dynamic Media
I removed hard coded asset importing and made changes to the component
to set the image and video based on the game data.

Along with the above change I also moved my assets to the public folder
to keep things simple. Typically, I would upload them to a S3 bucket or
create a component to dynamically load the images.