Production Build Base Url
Updated the `vite.config.js` file to set a base url when a production
build is created. This was necessary for my use case since I'm pushing
this to a subfolder in a statically hosted website in a S3 bucket.

Video and image sources were also updated to use the base url.